refactor(package): refactor package module in core to align with right folders
