+++ packages/core/src/package/SfpPackageBuilder.ts
@@ -2,8 +2,8 @@ import ApexTypeFetcher from '../apex/parser/ApexTypeFetcher';
 import ProjectConfig from '../project/ProjectConfig';
 import SfpPackageContentGenerator from './generators/SfpPackageContentGenerator';
 import SourceToMDAPIConvertor from './packageFormatConvertors/SourceToMDAPIConvertor';
-import PackageManifest from './PackageManifest';
-import MetadataCount from './MetadataCount';
+import PackageManifest from './components/PackageManifest';
+import MetadataCount from './components/MetadataCount';
 import SFPLogger, { Logger, LoggerLevel } from '@dxatscale/sfp-logger';
 import * as fs from 'fs-extra';
 import path from 'path';
@@ -18,7 +18,7 @@ import CreateSourcePackageImpl from './packageCreators/CreateSourcePackageImpl';
 import CreateDataPackageImpl from './packageCreators/CreateDataPackageImpl';
 import ImpactedApexTestClassFetcher from '../apextest/ImpactedApexTestClassFetcher';
 import * as rimraf from 'rimraf';
-import PackageToComponent from './PackageToComponent';
+import PackageToComponent from './components/PackageToComponent';
 import lodash = require('lodash');
 import { EOL } from 'os';
 import PackageVersionUpdater from './version/PackageVersionUpdater';
