+++ packages/core/src/package/components/MetadataCount.ts
