+++ packages/core/src/package/components/PackageManifest.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import * as fs from 'fs-extra';
-import { ApexClasses } from './SfpPackage';
-import xml2json from '../utils/xml2json';
+import { ApexClasses } from '../SfpPackage';
+import xml2json from '../../utils/xml2json';
 const xml2js = require('xml2js');
 
 export default class PackageManifest {
